+++ CHANGELOG.md
@@ -28,7 +28,7 @@
 
 ## 1.0.6
 
-- Prepare for 1.0.0 version of sensors and package_info. ([dart_lsc](http://github.com/amirh/dart_lsc))
+- Prepare for 1.0.0 version of sensors and package_info. ([dart_lsc](https://github.com/amirh/dart_lsc))
 
 ## 1.0.5